[PM-19577] Move OSLogErrorReporter to BitwardenKit and fix Authenticator build
