+++ Bitwarden/Application/AppDelegate.swift
@@ -1,3 +1,4 @@
+import BitwardenKit
 import BitwardenShared
 import UIKit
