+++ BitwardenAutoFillExtension/CredentialProviderViewController.swift
@@ -1,4 +1,5 @@
 import AuthenticationServices
+import BitwardenKit
 import BitwardenSdk
 import BitwardenShared
 import Combine
