+++ BitwardenKit/Core/Platform/Utilities/OSLogErrorReporter.swift
@@ -1,10 +1,12 @@
-import BitwardenKit
 import OSLog
 
 /// An `ErrorReporter` that logs non-fatal errors to the console via OSLog.
 ///
 public final class OSLogErrorReporter: ErrorReporter {
     // MARK: Properties
+
+    /// A list of additional loggers that errors will be logged to.
+    private var additionalLoggers: [any BitwardenLogger] = []
 
     /// The logger instance to log local messages.
     let logger = Logger(subsystem: Bundle.main.bundleIdentifier!, category: "ErrorReporter")
@@ -21,8 +23,17 @@ public final class OSLogErrorReporter: ErrorReporter {
 
     // MARK: ErrorReporter
 
+    public func add(logger: any BitwardenLogger) {
+        additionalLoggers.append(logger)
+    }
+
     public func log(error: Error) {
         logger.error("Error: \(error)")
+
+        let callStack = Thread.callStackSymbols.joined(separator: "\n")
+        for logger in additionalLoggers {
+            logger.log("Error: \(error as NSError)\n\(callStack)")
+        }
 
         // Don't crash for networking related errors.
         guard !error.isNetworkingError else { return }
